test: rename test dummy to DummyData
