+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/datasource/account/AccountPostFavoriteRemoteDataSourceTest.kt
@@ -1,7 +1,7 @@
 package com.waffiq.bazz_movies.core.network.data.remote.datasource.account
 
 import com.waffiq.bazz_movies.core.network.testutils.BaseMediaDataSourceTest
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.postResponseSuccessDump
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.postResponseSuccessDump
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testError404Response
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testErrorResponse
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testGeneralExceptionResponse
