+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/datasource/auth/AuthAccountDetailsRemoteDataSourceTest.kt
@@ -1,7 +1,7 @@
 package com.waffiq.bazz_movies.core.network.data.remote.datasource.auth
 
 import com.waffiq.bazz_movies.core.network.testutils.BaseAuthDataSourceTest
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.accountDetailsResponse
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.accountDetailsResponse
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testError404Response
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testErrorResponse
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testGeneralExceptionResponse
