+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/datasource/movie/MovieDetailRemoteDataSourceTest.kt
@@ -1,8 +1,8 @@
 package com.waffiq.bazz_movies.core.network.data.remote.datasource.movie
 
 import com.waffiq.bazz_movies.core.network.testutils.BaseMediaDataSourceTest
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.detailCollectionsResponse
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.detailMovieResponseDump
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.detailCollectionsResponse
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.detailMovieResponseDump
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testError404Response
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testErrorResponse
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testGeneralExceptionResponse
