+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/datasource/movie/MoviePostRateRemoteDataSourceTest.kt
@@ -2,7 +2,7 @@ package com.waffiq.bazz_movies.core.network.data.remote.datasource.movie
 
 import com.waffiq.bazz_movies.core.network.data.remote.models.RatingRequest
 import com.waffiq.bazz_movies.core.network.testutils.BaseMediaDataSourceTest
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.ratePostResponseSuccessDump
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.ratePostResponseSuccessDump
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testError404Response
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testErrorResponse
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testGeneralExceptionResponse
