+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/datasource/movie/MovieRemoteDataSourceTest.kt
@@ -2,11 +2,11 @@ package com.waffiq.bazz_movies.core.network.data.remote.datasource.movie
 
 import com.waffiq.bazz_movies.core.network.data.remote.pagingsources.GenericPagingSource
 import com.waffiq.bazz_movies.core.network.testutils.BaseMediaDataSourceTest
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.movieDump1
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.movieDump2
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.movieDump3
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.movieDump4
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.movieDump5
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.movieDump1
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.movieDump2
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.movieDump3
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.movieDump4
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.movieDump5
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.defaultMediaResponse
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testPagingFlow
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testPagingSource
