+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/datasource/tv/TvStateRemoteDataSourceTest.kt
@@ -1,7 +1,7 @@
 package com.waffiq.bazz_movies.core.network.data.remote.datasource.tv
 
 import com.waffiq.bazz_movies.core.network.testutils.BaseMediaDataSourceTest
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.statedResponseDump2
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.statedResponseDump2
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testError404Response
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testErrorResponse
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testGeneralExceptionResponse
