+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/responses/countryip/CountryIPResponseTest.kt
@@ -1,6 +1,6 @@
 package com.waffiq.bazz_movies.core.network.data.remote.responses.countryip
 
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.countryIPResponseDump
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.countryIPResponseDump
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Test
