+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/responses/tmdb/MediaResponseItemTest.kt
@@ -1,7 +1,7 @@
 package com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb
 
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.movieDump1
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.tvShowDump1
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.movieDump1
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.tvShowDump1
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Test
