+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/responses/tmdb/account/AccountDetailsResponseTest.kt
@@ -1,6 +1,6 @@
 package com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.account
 
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.accountDetailsResponse
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.accountDetailsResponse
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
