+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/responses/tmdb/media/castcrew/MediaCastItemResponseTest.kt
@@ -1,6 +1,6 @@
 package com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.media.castcrew
 
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.mediaCastItemResponseDump
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.mediaCastItemResponseDump
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
