+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/responses/tmdb/post/PostResponseTest.kt
@@ -1,6 +1,6 @@
 package com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.post
 
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.postResponseDump
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.postResponseDump
 import org.junit.Assert.assertEquals
 import org.junit.Assert.assertNull
 import org.junit.Assert.assertTrue
