+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/data/remote/responses/tmdb/state/MediaStateResponseTest.kt
@@ -1,7 +1,7 @@
 package com.waffiq.bazz_movies.core.network.data.remote.responses.tmdb.state
 
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.mediaStateResponse
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.ratedResponse
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.mediaStateResponse
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.ratedResponse
 import org.junit.Test
 import kotlin.test.assertEquals
 import kotlin.test.assertTrue
