+++ core/network/src/test/kotlin/com/waffiq/bazz_movies/core/network/testutils/BaseMediaDataSourceTest.kt
@@ -30,7 +30,7 @@ import com.waffiq.bazz_movies.core.network.data.remote.retrofit.services.PersonA
 import com.waffiq.bazz_movies.core.network.data.remote.retrofit.services.SearchApiService
 import com.waffiq.bazz_movies.core.network.data.remote.retrofit.services.TrendingApiService
 import com.waffiq.bazz_movies.core.network.data.remote.retrofit.services.TvApiService
-import com.waffiq.bazz_movies.core.network.testutils.DataDumpManager.tvShowDump1
+import com.waffiq.bazz_movies.core.network.testutils.DummyData.tvShowDump1
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.defaultMediaResponse
 import com.waffiq.bazz_movies.core.network.testutils.TestHelper.testPagingFlow
 import com.waffiq.bazz_movies.core.test.MainDispatcherRule
